[Init] Intialize the basic interface of explainition.
